backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic[email]==2.6.4
pydantic-settings==2.2.1

# PDF parsing
pdfplumber==0.11.0
PyMuPDF==1.24.0

# NLP
spacy==3.7.4
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl

# Embeddings
sentence-transformers==2.7.0
numpy==1.26.4

# Supabase
supabase==2.4.3
python-jose[cryptography]==3.3.0

# HTTP client
httpx==0.27.0

# Utilities
python-dotenv==1.0.1
aiofiles==23.2.1
